refactor: reorganizar paquetes de componentes para mejorar la estructura del proyecto
