+++ app/src/main/java/me/rogerroca/supernow/core/ui/headers/ScreenHeader.kt
@@ -1,14 +1,14 @@
-package me.rogerroca.supernow.core.ui.components
+package me.rogerroca.supernow.core.ui.headers
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import me.rogerroca.supernow.core.ui.misc.UserAvatar
 import me.rogerroca.supernow.core.ui.theme.AppTheme
 import me.rogerroca.supernow.core.ui.theme.AppTypography
