+++ app/src/main/java/me/rogerroca/supernow/features/products/ui/CardButtonGroup.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import me.rogerroca.supernow.R
-import me.rogerroca.supernow.core.ui.components.buttons.AppIconButton
-import me.rogerroca.supernow.core.ui.components.buttons.AppOutlinedIconButton
-import me.rogerroca.supernow.core.ui.components.textfields.NumberTextField
+import me.rogerroca.supernow.core.ui.buttons.AppIconButton
+import me.rogerroca.supernow.core.ui.buttons.AppOutlinedIconButton
+import me.rogerroca.supernow.core.ui.textfields.NumberTextField
 import me.rogerroca.supernow.core.ui.theme.AppTheme
 import me.rogerroca.supernow.core.ui.theme.buttons.narrowDefaultIconButtonModifier
 import me.rogerroca.supernow.core.ui.theme.buttons.primaryIconButtonColors
